# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for EventHub.",
        "title": "EventHub (GoHub) Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Tran Phuoc Anh Quoc",
            "email": "anhquoctpdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE"
        },
        "version": "OAS 3.0"
    },
    "paths": {
...
        "/api/v1/statistics/customer-conversion-rate": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the customer conversion rate",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the customer conversion rate",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/customer-retention-rate": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the customer retention rate",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the customer retention rate",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/customer-segmentation-rate": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the customer segmentation rate",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the customer segmentation rate",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/event-period-revenue": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the event period revenue",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the Event Period Revenue",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/event-reviews-by-customer": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the event reviews by customer",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the event reviews by customer",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/event-total-statistic": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the event total statistic",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the event total statistic",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/overview-sale-statistic": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the overview sale statistic",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the overview sale statistic",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/overview-total-statistic": {
            "get": {
                "description": "Fetches a paginated list of reviews based on the provided filter parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Retrieve the overview total statistic",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the overview total statistic",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/get-created-tickets": {
            "get": {
                "description": "Fetches the details of a specific expense based on the provided expense ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Retrieve an tickets by created",
                "responses": {
                    "200": {
                        "description": "Category created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not have the required permissions",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found - Category with the specified ID not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An error occurred while processing the request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}